apps/issue-2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-render"
version = "0.1.0"
description = "Render pixel matrix JSON into GIF or PNG files."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.4.0,<12.0.0"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "build>=1.2.0,<2.0.0",
]

[project.scripts]
pixel-render = "pixel_render.cli:run"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--basetemp=.pytest_tmp"
tmp_path_retention_policy = "none"
tmp_path_retention_count = 0
